[REEF-1528] Refactor dependency between yarn runtime and reef-webserver projects
This addressed the issue by
  * Reverting teh dependency between reef-webserver and reef-runtime
  * Fixing tests
  * Adding Empty implementation for the HttpServer
  * Changing default implementation of TrackingURLProvider to HttpTrackingURLProvider
  * Changing HttpTrackingURLProvider to return empty URL when http server is not configured (DefaultHttpServerImpl is configured)

JIRA:
  [REEF-1528](https://issues.apache.org/jira/browse/REEF-1528)
