+++ lang/java/reef-webserver/src/main/java/org/apache/reef/webserver/HttpServer.java
@@ -18,9 +18,12 @@
  */
 package org.apache.reef.webserver;
 
+import org.apache.reef.tang.annotations.DefaultImplementation;
+
 /**
  * HttpServer interface.
  */
+@DefaultImplementation(DefaultHttpServerImpl.class)
 public interface HttpServer {
 
   /**
